fix: 안 쓰이는 코드 제거
